split out http and sse capabilities
